server/requirements.txt
Flask~=3.0.3
python-dotenv~=1.0.1
langchain-community~=0.3.1
langchain-text-splitters~=0.3.0
langchain-huggingface~=0.1.0
google-generativeai~=0.8.2
pinecone~=5.3.1